Throughput optimisation.
Previously the proxy process would send all updated append entries
immediately if they were different from the previous set. This probably
caused many duplicate entries to be sent as well as many more messages.
Now the called has to pass an `IsUrgent` flag for the proxy to send
immediately. Currently this is only done as an effect of a newly
appended command so that new command latency isn't affected.

[#137888653]