Pipeline commands.
When a leader receives a command each peer's next_index is
optimistically incremented the the next index after the last included
entry. The assumption here is that the append entry rpc will succeed and
thus reduce the number of duplicate entries that are sent to the client
should the reply be delayed. This works better with RA's more
asyncronous (fake rpc) approach.

[#137888653 ]