fsync periodically on leader.
Leaders fsync every 10ms instead of after each command. This yields a good
throughput improvement at the expense of latency when leader is required
for consensus (e.g. in a 3 node cluster and one follower is
unavailable).

Also implemented simple cache for the last written entry in ra_log_file.

[#137888653]