ra_node: Reset `leader_id` when starting an election
Before that, we kept the old `leader_id` value around. If the
same leader is elected, the `leader_id` remained the same:
this prevented the `pending_commands` from being processed in
`ra_node_proc:follower_leader_change()`.

In the `ra_SUITE:node_recovery()` testcase, this caused a deadlock
because the synchronous `gen_statem:call()` in the second
`ra:send_and_await_consensus()` never got a reply.