WAL rollover
WAL can be configured to roll over before it reaches a certain size in
bytes. It then notifies the table writer process that the current
mem_tables need to be flushed to disk. Whilst flushing is taking place
in the background it moves the table lookups to an old mem_table ets
table.