Fix WAL bug
If a follower issued the following valid write requests:

WRITE Idx 5 Term 1
WRITE Idx 4 Term 2
WRITE Idx 5 Term 4

When the follower process then recieved the {written, 5} event is it not
clear if that refers to the first or second request to write index 5. As
overwriting the log is valid for a follower this could have resulted in
Idx 5 in term 2 having been confirmend back the leader before it had
actually been fully written.

[#152041636]