Add ra_log_file_segment
Implements a segment of a ra log.
The mem_tables created during the
lifetime of a wal file are flushed to ra node specific storage using a
segmented approach.

It is somewhat similar to an SSTable in a LSM tree based storage engine
as the keys are implicitly sorted (by virtue of being part of an indexed
log) and they are never compacted.
Each segment stores a range of log indexes. Once a
ra node's release_cursor moves beying the last index of a segment it can
simply be deleted which reduces the write-amplification compared to
using an LSM tree based storage engine directly.

[#152041636]