Implement log file read sequence.
The sequential log needs to be reconstructed from data stored in several
places.
Firstly we check the ra_log_file cache map which holds any in-flight
entries that have not yet been confirmed written by the WAL. Next we
query the ra_log_open_mem_tables to see if the range requested is
contained in the current open mem_table. After that we check the closed
mem tables. There could be several closed mem tables if the segment
writer has a processing backlog although typically there should be none.
Lastly we query all available on-disk segments in order to see if they can serve
the request.

[#152041636]