WAL truncating writes
After a snapshot we know the next WAL write may not be sequential and
thus need a means to signal to the wal that we need to truncate the
current mem_table from this point.

This info is persisted as the first bit in the WAL record header so that
we can restore this information after recovery.

[#152041636]