Improve WAL recovery logic
When the WAL starts it first waits for the segment writer to either
process whatever is in flight or confirm it is down. Then it proceeds to
process all remaining wal files into mem tables using a temporary mem
table lookup table to avoid interfering with the current open mem
tables. When all mem tables have been added to closed mem tables the
current open tables are deleted (as have been re-written already) and
the open mem table lookup table is cleared.

[#152041636]