Change supervision tree
As there is no means of re-sending mem_tables from the WAL to the
segment writer the supervision tree now is such that the WAL depends on
the segment writer but not the other way around (the WAL has a recover
procedure). The supervision tree also reflects that the WAL and
segwriter both depend on ra_log_file_ets.

What is missing is the dependency ra_nodes have on ra_log_file_ets.

[#152041636]