Emit state for release cursor
Rather than just using the intial state the release cursor effect now
includes the machine state to use for the snapshot. This could be the
empty state or a dehydrated version of the state at that time including
only the information necessary to restore the current state from the
snapshot and remaining log entries.

This does mean the state machine may need to track this state as is done
in ra_fifo which could result in a memory use penalty. Structural
sharing should help with this in the general case.

[#152041636]