Implement write strategies in the wal
Three different strategies are provided to allow users to select if they
want to optimise for speed or memory use.

`no_delay` will write each entry as soon as possible so that the data can
be reclaimed.

`delay_writes` will buffer all data and write it in one go at completion
of a write batch. This will reduce the number of syscalls but also means
that generated binary data may get promoted to the old heal and take
longer to be reclaimed.

`delay_writes_sync`  works like `delay_writes` but rather than doing a
file:write/2 + file:sync/1 it tries to open the wal files using
synchronous IO (O_SYNC). This is the fastest option.

[#152041636]