Replace dets with custom fixed size data store.
dets tables are accessed and written through a dedicated process. This
is fine for the commit_index and voted_for keys as they are written
rarely, however, if we wanted to provide an option to persiste the
last_applied index this will be done much more frequently and it would hurt
performance if we had to constantly call into another process.

[#152041636]