Introduce a local unique identity for ra processes.
Within an erlang node a ra process now has a unique binary identity that
is used to create a storage directory as well as to provide an identity
for wal and segment writer interactions. A ra_directory module has been
introduced to provide registration and lookup from the binary id to the
current pid and registered name.

Peer to peer interactions are still performed using the more efficient
{atom(), node()} tuple and this is what provide a cluster identity.

[#154472158]