Distinquish between different process down events
If a process is detected as 'DOWN' with no connect we no longer remove
them as a consumer or enqueuer in ra_fifo and instead just mark them as
suspect and begin monitoring the status of the node their on. When that
nodes comes back with a `nodeup` we re-issue the process monitors to
detect their current status.