Re-introduce a cluster id to identify the cluster.
Clusters will need a common id for things like local cached leader lookup and
other coordination. This does mean there are 3 separate identifier that
all need to be configured:

UId - the local identifier that unique identifies a node inside an
erlang node.
NodeId - a ra node's addressable identity inside a ra cluster
ClusterId - the id of the cluster as a whole.