Implement pre_vote and shutdown sequence.
Shutdown is a consensus action. The leader replicates a termination
command. When a follower receives it it termates itself and deletes all
it's data. The leader waits for the termination command to be replicated
to _all_ followers before doing the same.

To make this work well it required the pre_vote phase to be implemented.
See §6 in raft thesis.