Auto increase ra_fifo_client max pending window
When adding a consumer and the num pending value is greater than the max
pending window it is possible to hit the flow control limit very
quickly when a large batch of messages that are quickly settled is
deliveryed.
To avoid this we now auto increase the window with the number of allowed
in-flight messages. This does mean that the window of uncertaintly w.r.t
these messages is larger but for now this is a trade-off we have to
make.