WAL improvements
Make wal configuration more consistent and document it.

Handle a recover scenario when the beam was killed whilst the WAL was
writing. If a parse error is encountered during wal recover we handle
this and discard all subsequent data as these writes were never
confirmed back to the caller.

Introduce ra:start_or_restart_cluster/3.