Reduce unnecessary segment writes
Allow the segments writer to skip writing entries that preceed the last
snapshot for a process. This is done by a new shared ETS table that the
ra_node records it's current snapshot index and the segment writer skips
all entries lower than this.

Also changed to periodically take snapshots even if no segments can be
cleared up in order to trigger the above behaviour.