Fix ra_log:take bug
If a ra_log received an update to a segment it as already read from it
would never re-read the index of that segment and silently drop entries
from the take operation. This fixes that by always closing any open
segmenents that have been updated. Although this may result in the file
almost immediately being open again it can also help with keeping file
descriptor use down.

Also added a few temporary assertions for this bug.