Move use calculation to ra_fifo
This did never belong inside core ra and has been moved to the
ra_fifo statemachine instead using an optional new "aux" machine api.

Aux allows ra to maintain state on behalf of the machine that isn't part
of the consistent state machine itelf. Operations can be invoked by
emitting an aux effect with a command term that is then passed to the
handle_aux function with the aux state, the machine state and the log.
Any interactions are allowed here even reading from the log (although
writing would be weird).