Limit file handle use by ra nodes
This enforces a limit on the number of simultaneous segmenst (and thus
file handles) a ra node can have open at any one time. It uses a fixed
size lru cache to maintain the open segments and closes the evicted
segments if the limit is reached.

Leaders set a limit equal to the cluster size and followers use a limit
of 1.

[#157003974]