Make ra machine apply return type more consistent
Instead of either returning a tuple of arity 2 or 3 with 3 containing a
reply value the ra machine apply function now always need to return a
value for every applied command. This value will be returned either as
the return value in ra:process_command/2 or with the correlation id when
using correlated pipelined requests.