send_msg effect sends plain message by default
Instead of wrapping a message sent in response to a send_msg effect in a
ra_event this changes it to a plain message send. The option to wrap it
in a ra_event is still there by passing the ra_event option with the
send_msg effect. Also added another option to wrap the message up as a
$gen_cast.

Refactored ra_log:read_cons to avoid using file:pread/2 as this involves
a lists:reverse/1 at the end which could potentially be expensive. This
was detected by the reduction counter limit in the ra_log_2 test suite
triggering when upgrading to OTP21.1. In OTP21.1 a bug where
lists:reverse/1|2 didn't bump the reductions enough was fixed which
resulted in significantly higher reduction counts.