Change ra_machine apply/4 into apply/3
This simplifies the ra_machine apply behaviour function not to take a
list of effects and necessitate that any further effects are cons-ed to
the existing list. Instead implemetors just return a list of effects in
execution order.

Also extended and amended the return type to avoid creating unnecessary
lists and tuple slots.