Pipelining improvements
Whena  follower slips behind and there isn't a natural ingress of
commands it may take a long time to catch up as the only message
triggering pipelining is the append entry reply messages which could
take several ms to come back. If pipelining is possible after creating
append entries rpcs we emit a next_event effect to force it for selected
situations (to avoid flooding the server with unnecessary pipeline
requests).