A function to execute a query on the (reasonably) current leader.
Works similar to local_query, but redirects the request to the
current leader.
There is no guarantee that the leader will bot be superceeded by
another node or is not already.
There is no guarantee that the state is the most up-to-date.

This query may return stale results.

The purpose of this query is to return slightly up-to date data,
as leader applies commands before followers.