Add log and timer effects
The timer effect asks the Ra leader to maintain a single relative timer
and commit a `timeout` command when it triggers.

The log effect allows the effect system to read and entry from the log
and turn it into an effect. Useful for when the machine doesn't want to
keep large data in memory but instead prefers to read it from the log
when needed for a side effect.