Implement versioning of state machines
This allows state machine implementations to support
deterministic upgrade of machine logic and state.

This introduces two additional optional ra_machine callbacks:

* version/0 to return the current state machine version integer
* which_module/1 to map a version to the module that implements it

The version is now persisted in the snapshot as well as in the leader
noop command that is commit at the beginning of each raft term which is
also how the switch from an old version to new is replicated.

[#164212395]