Make segment writer handle server deletion better.
Previously it would only handle the case where the ra server's directory
had been deleted _before_ it started flushing mem table entries to
segments. The segment writer would crash if the directory was deleted
whilst it was writing to segments if it then would try to open a new
segment.