Limit segment append memory use
This sets a limit (max_pending) on the number of writes the segment will
cache before it actually writes it to the file using file:pwrite/2. This
will ensure it doesn't create insanely large lists in memory and should
keep peak memory use under better control when mem tables are flushed to
disk.

Also change the tick_timeout code to base it's decision on whether to
garbage collect on whether the last applied index has moved or not
instead of using the last written. This is better as the last applied
can change well after the last index was written.