Adjust leader tracking logic
Leader tracking in Ra is done using two bits of state. The leader_id in
the ra_server and leader_monitor in the ra_server_proc. The
leader_monitor will be cleared upon receiving a DOWN notification from
the leader monitor and is used when deciding whether to cache incoming
commands until a leader is known. The leader_id is updated for followers
based on incoming append_entries_rpc messages. Prevously the code that
triggered ra to set up another leader monitor and registering the leader
node with aten etc was only done if the leader_id changed. Now it takes
the state of the leader_monitor into account to better handle the case
when a leader is disconnected, then comes back and is still a leader.