Handle duplicated entries in WAL in ra_dbg
Followers can overwrite somes entries, making the index going back in
the WAL. This is an implementation details, but this can make the replay
fail because the state machine will see the same entries twice, making
the state potentially inconsistent. This commit keeps only the latest
entry if some duplicated index shows up.