Fix segment writer bug
If the beam process is killed during segment flushing there is a chance
it will create an empty segment file. If the segment writer encounters
this file it (prior to this fix) handled the error as if the server
folder was no longer there. This would cause it to incorrectly delete
the WAL without flushing all entries to disk!

This fix handles this case and deletes the empty segment file when
found. It also simplifies the directory does not exist scenario.