Do not stash membership changes that can't be appended
Raft only supports one cluster change at a time. Previously we'd stash
any changes that came in whilst another was in progress and they would
later be applied after the current one was committed. This could result
in very hard to reason about behavour from the outside which is why this
commit removes it and provides more detailed error messages for the
ra:add_member|remove_member commands.