Fix integer overflow bug in segment index
The file offset was for some inexplicable reason recorded as a 32 bit
integer effectively limiting max file size to ~4GB. With large Ra
commands and large segment max entries configured it is fairly easy to
go beyond this causing data in the segment to be overwritten.

This change stores the offset in 64 bit but remains backwards compatible
by incrementing the segment file version and handling this in the code.