Optimise segment reads by implementing read ahead cache.
This cache is only enabled (access_mode=sequential) during
recovery as this is the only time the log read access pattern
is guaranteed to be sequential. After recovery it reverts back
to random mode which does not read ahead.

We do not use the build it `read_ahead` option that you can
pass to file:open/2 as this does not work with file:pread
which is what the log segment requires.