ra: Add support for local queries in `members/{1,2}`
If the (single) server ID is passed as `{local, ServerId}`, the
specified server handles the query, regardless of its role (leader or
follower). If the queried node is not the leader, it might return a possibly
out-of-date information.

This differs from the existing default behavior of `members/{1,2}` where
the request would be redirected to the leader, returning the up-to-date
consistent information.

This new option is useful if the caller can't wait for a leader to
appear and accepts to get stalled data.