ra_log_wal: Support writes larger than WAL max size
Initially, MaxWalSize was a hard limit for the file size: if FileSize +
DataSize went over that limit, we would use a new file. This was an
issue when DataSize was larger than the limit alone.

The chosen solution is to only consider the current file size in the
calculation. It means that after DataSize bytes are written, the file
will be larger than the configured maximum size. But at least it will
accept Ra commands larger than the max WAL size.

Before this change, ra_log_wal would enter an infinite loop opening new
files because the large command would not fit any of them.