Add two less safe wal write strategies.
1. sync_after_notify: this write strategy will send confirmation notifications
back to writing processes after writing but  _before_ initiating the fsync operation.

This should improve latency for Ra machines that do not need high throughput.

2. Sync Method: none: This sync method implements a noop for the
sync operation effectively disabling fsync apart from when closing the
file. Only to be used when the disk used provide additional guarantees. E.g.
when the disk has batteries and can ensure data is synced in case of a power
outage.