Fix bug when a log read is partially covered by the cache
A refactor must have accedentally missed to include any entries read
from the cache in the read result. This would have resulted in leader
creating gaps in the log replication messages causing followers to have
to ask the leader to back up. Everything still works it just does a bit
of unnecessary work.