ra_server: Export `command_priority()` type
Before this patch, the type was defined privalitely as
`ra_command_priority()` in the `ra_server_proc` module. The `ra` module
hard-coded the possible values of this type as well.

Now the type is defined once in `ra_server` and exported. This also
allows users of Ra to use it.