Use the correct ra sha in the oci workflow
From the github actions docs:
Note that GITHUB_SHA for this event is the last merge commit of the
pull request merge branch. If you want to get the commit ID for the
last commit to the head branch of the pull request, use
github.event.pull_request.head.sha instead.