Make first server trigger leader election
Previously, a local Ra server triggered leader election when starting a
new cluster.

This commit makes the first server in the list of servers to trigger
leader election.

This commit is necessary for quorum queue-leader-locator set to
'random' or 'least-leaders' where leaders can be non-local to the node
where ra:start_cluster/2,3,4 is called.