Handle machine versions in snapshot installs
When a followe with a lower effective machine version receives a
snapshot with a higher version (that it still understands) we need
to treat this similarly to the "noop" command and ensure that from that point
on the server uses the correct effective machine version.

Also ignore any snapshot_installs for machine versions a follower
cannot understand. At some point we may allow followers to accept
snapshots they do not yet understand like is allowed for normal
log entries but that involves keeping multiple snapshot versions on disk
which is moderately complext for something that only very rarely will be needed.