Wait for Ra servers exit in `delete_cluster/{1,2}`
So far, when the `delete_cluster/{1,2}` function returned, it meant that
the request to stop the cluster was committed to the Ra log and applied.
However, the actual stop of the Ra servers and the related supervision
tree was handled asynchronously.

This patch uses Erlang monitors to watch the exit of the Ra servers.
This way, we ensure that when the function returns, the Ra servers are
really gone.