ra: Wait for servers exit in `delete_cluster/3` if requested
The caller can pass a flag to `delete_clsuter/3` to indicate if he wants
the function to wait for servers to exit. He can also set the
application environment variable
`wait_for_servers_exit_in_delete_cluster` to define the global behavior.

The default behavior is the same as the historical one, meaning the
function doesn't wait for servers exit.