Remove lists:reverse/1 call in ra_log_wal
The list of inserts for a given server / wal batch was always reversed
before passed to ets:insert. This is almost never needed as it is very
rare for an overwrite to happen in the same wal batch and just results
in unnecessary cpu and allocations.

It is now replaced with a check + lists:dropwhile if an overwrite in the same
batch is detected.