WAL: handle corrupt last record
The last record in a WAL file can be a partial write and thus fail
the CRC check when the WAL is pre-allocated. In this case we check
that the failed CRC check record is the last record in the WAL.
When this is the case the record is dropped and recovery
resumes.