ra_SUITE: Add a test to demonstrate race condition in follower_leader_change()
The goal is to send a command to a follower in the middle of an
election. Before any fix is committed, there is a great chance that the
follower will redirect pending commands to the previous exited leader,
causing the callers to time out or get a `noproc` error.

The reason is that the follower will call `follower_leader_change()`
several times before it leaves its "follower" state for e.g. "pre_vote",
if it receives other events before the change of state.