ra_server_proc: Clear leader ID after receiving a DOWN event
If the DOWN reason is something else than `noconnection`, we know that
the leader process is really gone. We want to clear the leader ID we
know at this point, while a new election is running.

This is to make sure that `follower_leader_change()` won't create a
useless monitor and more importantly, it won't redirect pending
commands to that old leader. This would cause the commands callers to
get a `noproc' error or a timeout.