Clear leader ID after receiving a DOWN event in `ra_server_proc` (#291)
* ra_SUITE: Add a test to demonstrate race condition in follower_leader_change()

The goal is to send a command to a follower in the middle of an
election. Before any fix is committed, there is a great chance that the
follower will redirect pending commands to the previous exited leader,
causing the callers to time out or get a `noproc` error.

The reason is that the follower will call `follower_leader_change()`
several times before it leaves its "follower" state for e.g. "pre_vote",
if it receives other events before the change of state.

* ra_server_proc: Clear leader ID after receiving a DOWN event

If the DOWN reason is something else than `noconnection`, we know that
the leader process is really gone. We want to clear the leader ID we
know at this point, while a new election is running.

This is to make sure that `follower_leader_change()` won't create a
useless monitor and more importantly, it won't redirect pending
commands to that old leader. This would cause the commands callers to
get a `noproc' error or a timeout.

* ra_SUITE: Enable logging in `send_command_to_follower_during_election`

This might help us understand why it fails in CI.

* test reliability

Co-authored-by: Karl Nilsson <kjnilsson@gmail.com>