Reply and notify _after_ apply effects have been executed.
This changes the order of notify and reply effects such that any effects
emitted by the application of a command are executed before the caller
gets a notication or a reply. This ensures that things like updating
and ETS table has been done by the time the caller gets a reply.