Revert notify effect processing order done in 773aa85e8ba79c2aed2f4eb766c97ff525d3e2ce
This introduces additional write latency that is dependent on batch size
and other effect processing times. Also it doesn't provide strong
enough consistency guarantees as it tells you nothing about the
effect processing on any nodes but the leader. It is better to assume
eventual consistency when it comes to effect processing anyway and
design systems accordingly.

fix tests