ra_system: Fix stop/1 when the `ra` application is stopped
If the `ra` application is stopped, the `ra_systems_sup` supervisor is
not running. This leads to a `{noproc, _}` exit exception if we try to
stop a Ra system.

This exception should be caught to make sure that the function always
work and is idempotent. Indeed, the Ra system is already stopped if the
application is not running.