Various fixes to ra_leaderboard
Ensure that the members are updated when the membership
is modifiied.

Always delete leaderboard record in ra server terminate callback
to favour accuracy of leadership information over availability.

The idea is that when a record is not found in the leaderboard
the caller falls back to some other, potentially slower, means
of leader discovery.