Store checkpoints in `#ra_snapshot{}`
Snapshot state already holds nearly everything necessary for checkpoints.
We add a parameter and a field to the record for the checkpoint
directory and adjust some of the `ra_snapshot` API to differentiate
between snapshots and checkpoints.

An alternative approach to this would be to have a separate
`#ra_snapshot{}` state for snapshots and checkpoints. However we would
need to change `#ra_snapshot{}` somewhat substantially to support
multiple snapshots, even though there is only ever one useful snapshot.
It also makes it slightly harder to ensure that there is only one
checkpoint or snapshot being written at a given time, whereas this
strategy can continue to use the `pending` field.