Take checkpoints with the `{checkpoint,Idx,MacState}` effect
This updates the snapshotting codepath to accept a new parameter

    SnapKind :: snapshot | checkpoint

otherwise the snapshotting codepath can be almost entirely reused for
checkpointing.

We don't yet properly handle the `snapshot_written` event and properly
complete either the snapshot or the checkpoint depending on which was
written - that will be covered in the child commit.