ra_log: Handle snapshot kind in `snapshot_written` event
`snapshot_written` with a `ra_snapshot:kind()` of `snapshot` retains
the same behavior as before checkpoints. We add a clause to handle
`checkpoint`s though which adds the checkpoint to the `#ra_snapshot{}`
state and thins out any surplus checkpoints.

When we write a regular `snapshot` kind, we also remove any checkpoints
older than the snapshot's index. These older checkpoints have no use
once there's a snapshot newer than them: recovery would only be slower
and there's no point in promoting a checkpoint for a state older than
the current snapshot - it would be a no-op.