Add a `{release_cursor,Idx}` effect to promote checkpoints
This effect allows you to turn a checkpoint into a snapshot. This is a
useful operation for Ra machine's like `rabbit_fifo` (quorum queues)
which read information out of the log. `rabbit_fifo` may only snapshot
up to an index where all enqueue/requeue commands before that index
have been dequeued because snapshotting deletes the enqueue/requeue
commands from the log. So `rabbit_fifo` needs a way to express that it
can snapshot up to a certain index which is not necessarily the current
index when it emits the `release_cursor` effect. Checkpoints contain
states at older indices which are perfect for this - they just need to
be moved from `DataDir/checkpoints/` to `DataDir/snapshots/`. Using a
file rename minimizes the I/O cost of promotion.