Recover from a checkpoint if available
This should greatly improve recovery time in degenerate cases for
machines like `rabbit_fifo`. `rabbit_fifo` only snapshots up to indices
where all prior enqueued messages have been consumed. So recovery time
is proportional to the number of outstanding enqueued messages unless
we can recover from a checkpoint. With this change and switching to
checkpoints in `rabbit_fifo`, recovery should be roughly constant-time.

The only part of the code necessary to change for this is
`ra_snapshot:recover/1`. That function would previously recover from
the snapshot if it existed but now it can check to see if there is a
more recent checkpoint. The rest of the machinery in `ra_log`
transparently takes care of recovering from the checkpoint's index.