Only fsync checkpoints before promotion
Snapshots must be fsync'd so that we can safely truncate the log.
There's no need to fsync every checkpoint though: we only care about
a checkpoint surviving its write to disk when we go to promote it.
So we can move the call to `file:sync/1` into the promotion task. By
promotion time it's possible that the file will already be synced.