Fix cluster recovery bug.
In Raft the cluster is changed when a cluster change command is
written to the log. Hence to correctly recover the current cluster
a state machine needs to scan until the end of the log on recovery
to ensure it picks up any observed changes.

Ra didn't do this and thus were vulnerable to pick up the wrong
cluster configuration (which ofc would affect subsequent leader elections).