Possibly avoid a rare race condition in the WAL
When the wal detects a new writer it creates a new ETS mem table
and inserts an initial record into the open mem table.. hrm.. table.

This is so that the open_mem_tables table can later be updated using
ets:update_element/3.

However the initial update was such that it indicated that it should
contain a single entry which isn't yet true. The entires are written
a little later during batch completion.

This is an attempt to avoid this condition by setting the "To"
field to -1 that of the "From" field.