When using multi statem call only retry for certain errors.
Instead of trying all server ids in the supplied list whenever an
error or timeout was encountered this change only retries when
certain errors such as noproc, nodedown etc is encountered.

If all errors result in a retry genuine API errors such as
{error, not_member} in reposnse to a ra:remove_member/2 interaction
would be hidden in the aggregate error that is returned once
all members have returned.