Return an error if a `call` is unsupported
[Why]
Currently, if a call to a Ra server is unsupported, the Ra server will
ignore the event and the call will eventually time out. This could
happen when Ra servers sport different versions of Ra for instance.

It would be nicer if the Ra server would reply immediatly with an error
to let the caller know about the actual problem.

[How]
The Ra server uses a `reply` effect in the catch-all clause.

The returned error is:

    {error, {unsupported_call, Call}}